refactor (setup build for module only)
