+++ .gitignore
@@ -1,4 +1,4 @@
 .DS_Store
 .idea/*
 node_modules/*
-dist/*
+**/dist/*
