+++ src/app/modules/jdnconvertiblecalendardateadapter/.npmignore
@@ -1,0 +1,2 @@
+.DS_Store
+adapter/*
