+++ src/app/modules/jdnconvertiblecalendardateadapter/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "jdnconvertiblecalendardateadapter",
+  "version": "0.0.2",
+  "description": "Implementation of Angular Material 2 DateAdapter for JDNConvertibleCalendar",
+  "types": "dist/index.d.ts",
+  "main": "dist/index.js",
+  "scripts": {
+    "prepublishOnly": "npm run prepare",
+    "prepare": "tsc",
+    "test": ""
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/dhlab-basel/JDNConvertibleCalendarDateAdapter.git"
+  },
+  "author": "Tobias Schweizer <t.schweizer@unibas.ch>, DHLab University of Basel (http://www.dhlab.unibas.ch)",
+  "license": "AGPL-3.0",
+  "bugs": {
+    "url": "https://github.com/dhlab-basel/JDNConvertibleCalendarDateAdapter/issues"
+  },
+  "homepage": "https://github.com/dhlab-basel/JDNConvertibleCalendarDateAdapter#readme",
+  "dependencies": {
+    "@angular/animations": "^5.1.0",
+    "@angular/cdk": "^5.0.1",
+    "@angular/common": "^5.1.0",
+    "@angular/compiler": "^5.1.0",
+    "@angular/core": "^5.1.0",
+    "@angular/forms": "^5.1.0",
+    "@angular/http": "^5.1.0",
+    "@angular/material": "^5.0.1",
+    "@angular/platform-browser": "^5.1.0",
+    "@angular/platform-browser-dynamic": "^5.1.0",
+    "@angular/platform-server": "^5.1.0",
+    "@angular/router": "^5.1.0",
+    "@types/core-js": "^0.9.44",
+    "angular-tree-component": "^6.1.0",
+    "jdnconvertiblecalendar": "0.0.1",
+    "rxjs": "^5.5.2",
+    "ts-helpers": "^1.1.2",
+    "zone.js": "^0.8.16"
+  }
+}
