+++ src/app/modules/jdnconvertiblecalendardateadapter/tsconfig.json
@@ -15,5 +15,8 @@
       "es2017",
       "dom"
     ]
-  }
+  },
+  "exclude": [
+    "./adapter/JDNConvertibleCalendarDateAdapter.spec.ts"
+  ]
 }
